lib/agents/workflow-template-library.ts
import { BaseAgent, AgentConfig } from './base';

export interface WorkflowTemplateLibraryConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class WorkflowTemplateLibraryAgent extends BaseAgent {
  private config: WorkflowTemplateLibraryConfig;

  constructor(config: WorkflowTemplateLibraryConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Workflow Template Library implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Workflow Template Library completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced workflow template library processing
    return { status: 'success', data };
  }
}

// FEATURE UPDATE: Workflow Template Library
// Date: 2024-08-12
import { BaseAgent, AgentConfig } from './base';

export interface WorkflowTemplateLibraryConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class WorkflowTemplateLibraryAgent extends BaseAgent {
  private config: WorkflowTemplateLibraryConfig;

  constructor(config: WorkflowTemplateLibraryConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Workflow Template Library implementat...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-08-20
import { BaseAgent, AgentConfig } from './base';

export interface WorkflowTemplateLibraryConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class WorkflowTemplateLibraryAgent extends BaseAgent {
  private config: WorkflowTemplateLibraryConfig;

  constructor(config: WorkflowTemplateLibraryConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Workflow Template Library implementat...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-10-24
import { BaseAgent, AgentConfig } from './base';

export interface WorkflowTemplateLibraryConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class WorkflowTemplateLibraryAgent extends BaseAgent {
  private config: WorkflowTemplateLibraryConfig;

  constructor(config: WorkflowTemplateLibraryConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Workflow Template Library implementat...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-10-31
import { BaseAgent, AgentConfig } from './base';

export interface WorkflowTemplateLibraryConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class WorkflowTemplateLibraryAgent extends BaseAgent {
  private config: WorkflowTemplateLibraryConfig;

  constructor(config: WorkflowTemplateLibraryConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Workflow Template Library implementat...
