__tests__/lib/agents/workflow-template-library.test.ts
import { WorkflowTemplateLibraryAgent } from '@/lib/agents/workflow-template-library';

describe('Workflow Template LibraryAgent', () => {
  let agent: WorkflowTemplateLibraryAgent;

  beforeEach(() => {
    agent = new WorkflowTemplateLibraryAgent({
      name: 'Test Workflow Template Library Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Workflow Template Library completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});

// FEATURE UPDATE: Workflow Template Library
// Date: 2024-08-12
import { WorkflowTemplateLibraryAgent } from '@/lib/agents/workflow-template-library';

describe('Workflow Template LibraryAgent', () => {
  let agent: WorkflowTemplateLibraryAgent;

  beforeEach(() => {
    agent = new WorkflowTemplateLibraryAgent({
      name: 'Test Workflow Template Library Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successf...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-08-20
import { WorkflowTemplateLibraryAgent } from '@/lib/agents/workflow-template-library';

describe('Workflow Template LibraryAgent', () => {
  let agent: WorkflowTemplateLibraryAgent;

  beforeEach(() => {
    agent = new WorkflowTemplateLibraryAgent({
      name: 'Test Workflow Template Library Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successf...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-10-24
import { WorkflowTemplateLibraryAgent } from '@/lib/agents/workflow-template-library';

describe('Workflow Template LibraryAgent', () => {
  let agent: WorkflowTemplateLibraryAgent;

  beforeEach(() => {
    agent = new WorkflowTemplateLibraryAgent({
      name: 'Test Workflow Template Library Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successf...


// FEATURE UPDATE: Workflow Template Library
// Date: 2024-10-31
import { WorkflowTemplateLibraryAgent } from '@/lib/agents/workflow-template-library';

describe('Workflow Template LibraryAgent', () => {
  let agent: WorkflowTemplateLibraryAgent;

  beforeEach(() => {
    agent = new WorkflowTemplateLibraryAgent({
      name: 'Test Workflow Template Library Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successf...
